pyproject.toml
[project]
name = "meraki-devices"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "icecream>=2.2.0",
    "meraki>=2.2.0",
    "python-decouple>=3.8",
]
